requirements.txt
streamlit
psycopg[binary]
pandas
python-dotenv
reportlab
